........
  r279113 | tilghman | 2010-07-23 13:56:04 -0500 (Fri, 23 Jul 2010) | 2 lines
  
  Silly 64-bit compilers (who uses 64-bit anyway?)
........


git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@279115 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ res/res_odbc.c
@@ -1479,8 +1479,8 @@ static odbc_status odbc_obj_connect(struct odbc_obj *obj)
 		ast_mutex_unlock(&obj->lock);
 		return ODBC_FAIL;
 	}
-	SQLSetConnectAttr(obj->con, SQL_LOGIN_TIMEOUT, (SQLPOINTER *) obj->parent->conntimeout, 0);
-	SQLSetConnectAttr(obj->con, SQL_ATTR_CONNECTION_TIMEOUT, (SQLPOINTER *) obj->parent->conntimeout, 0);
+	SQLSetConnectAttr(obj->con, SQL_LOGIN_TIMEOUT, (SQLPOINTER *)(long) obj->parent->conntimeout, 0);
+	SQLSetConnectAttr(obj->con, SQL_ATTR_CONNECTION_TIMEOUT, (SQLPOINTER *)(long) obj->parent->conntimeout, 0);
 #ifdef NEEDTRACE
 	SQLSetConnectAttr(obj->con, SQL_ATTR_TRACE, &enable, SQL_IS_INTEGER);
 	SQLSetConnectAttr(obj->con, SQL_ATTR_TRACEFILE, tracefile, strlen(tracefile));
